requirements.txt
fastapi
uvicorn
httpx
python-dotenv
gunicorn
pydantic-settings
